Move all components into a separate file
